requirements.txt
aiogram>=3.13,<4
openai>=1.55
python-dotenv>=1.0
asyncpg>=0.30
apscheduler>=3.10
sqlalchemy[asyncio]>=2.0
alembic>=1.13
pydantic>=2.0